feat!: check access in server actions

+++ src/actions/mutations.ts
@@ -1,8 +1,9 @@
 "use server";
 
+import { auth } from "@/auth";
 import { prisma } from "@/prisma";
 import { saltAndHashPassword } from "@/utils/auth/saltAndHashPassword";
-import { Course, CourseContent, Lecturer } from "@prisma/client";
+import { Course, CourseContent, Lecturer, UserRole } from "@prisma/client";
 import { User } from "next-auth";
 
 export type CreateUserInput = {
@@ -11,8 +12,32 @@ export type CreateUserInput = {
     password: string;
 };
 
+async function checkAdmin() {
+    const session = await auth();
+    // @ts-expect-error: session type istn changeable in the dependency
+    if (session?.user?.role == UserRole.ADMIN) {
+        return true;
+    }
+    return false;
+}
+
+async function checkLecturerAccess(courseId: number) {
+    const session = await auth();
+    // @ts-expect-error: session type istn changeable in the dependency
+    if (session?.user?.role == UserRole.ADMIN) return true;
+
+    const checkCourse: Course | null = await prisma.course.findUnique({
+        where: {
+            id: courseId
+        },
+    });
+    // @ts-expect-error: session type istn changeable in the dependency
+    if (checkCourse?.lecturerId == session?.user?.lecturerId && session?.user?.role == UserRole.LECTURER) return true;
+    return false;
+}
 
 export async function createCourse(courseData: Omit<Course, "id" | "createdAt" | "updatedAt">): Promise<Course> {
+    if (!await checkAdmin()) throw new Error("Not authorized");
     const resp = await prisma.course.create({
         data: courseData
     });
@@ -20,6 +45,7 @@ export async function createCourse(courseData: Omit<Course, "id" | "createdAt" |
 }
 
 export async function updateCourse(id: number, courseData: Omit<Course, "id" | "createdAt" | "updatedAt">): Promise<Course> {
+    if (!await checkLecturerAccess(id)) throw new Error("Not authorized");
     const resp = await prisma.course.update({
         where: {
             id
@@ -30,6 +56,7 @@ export async function updateCourse(id: number, courseData: Omit<Course, "id" | "
 }
 
 export async function deleteCourse(id: number): Promise<Course> {
+    if (!await checkLecturerAccess(id)) throw new Error("Not authorized");
     const resp = await prisma.course.delete({
         where: {
             id
@@ -39,6 +66,7 @@ export async function deleteCourse(id: number): Promise<Course> {
 }
 
 export async function createLecturer(lecturerData: Omit<Lecturer, "id" | "createdAt" | "updatedAt">): Promise<Lecturer> {
+    if (!await checkAdmin()) throw new Error("Not authorized");
     const resp = await prisma.lecturer.create({
         data: lecturerData
     });
@@ -46,6 +74,7 @@ export async function createLecturer(lecturerData: Omit<Lecturer, "id" | "create
 }
 
 export async function updateLecturer(id: number, lecturerData: Omit<Lecturer, "id" | "createdAt" | "updatedAt" | "courses">): Promise<Lecturer> {
+    if (!await checkAdmin()) throw new Error("Not authorized");
     const resp = await prisma.lecturer.update({
         where: {
             id
@@ -56,6 +85,7 @@ export async function updateLecturer(id: number, lecturerData: Omit<Lecturer, "i
 }
 
 export async function deleteLecturer(id: number): Promise<Lecturer> {
+    if (!await checkAdmin()) throw new Error("Not authorized");
     const resp = await prisma.lecturer.delete({
         where: {
             id
@@ -65,6 +95,7 @@ export async function deleteLecturer(id: number): Promise<Lecturer> {
 }
 
 export async function createCourseContent(courseId: number, contentData: Omit<CourseContent, "id" | "createdAt" | "updatedAt" | "courseId">): Promise<CourseContent> {
+    if (!await checkLecturerAccess(courseId)) throw new Error("Not authorized");
     const resp = await prisma.courseContent.create({
         data: {
             ...contentData,
@@ -74,20 +105,22 @@ export async function createCourseContent(courseId: number, contentData: Omit<Co
     return resp;
 }
 
-export async function updateCourseContent(id: number, contentData: Partial<Omit<CourseContent, "id" | "createdAt" | "updatedAt" | "courseId">>): Promise<CourseContent> {
+export async function updateCourseContent(courseId: number, contentData: Partial<Omit<CourseContent, "id" | "createdAt" | "updatedAt" | "courseId">>): Promise<CourseContent> {
+    if (!await checkLecturerAccess(courseId)) throw new Error("Not authorized");
     const resp = await prisma.courseContent.update({
         where: {
-            id
+            id: courseId
         },
         data: contentData
     });
     return resp;
 }
 
-export async function deleteCourseContent(id: number): Promise<CourseContent> {
+export async function deleteCourseContent(courseId: number): Promise<CourseContent> {
+    if (!await checkLecturerAccess(courseId)) throw new Error("Not authorized");
     const resp = await prisma.courseContent.delete({
         where: {
-            id
+            id: courseId
         }
     });
     return resp;
@@ -95,6 +128,7 @@ export async function deleteCourseContent(id: number): Promise<CourseContent> {
 
 /* eslint-disable @typescript-eslint/no-unused-vars */
 export async function createUser(user: CreateUserInput): Promise<Omit<User, "passwordHash">> {
+    if (!await checkAdmin()) throw new Error("Not authorized");
     const passwordHash = await saltAndHashPassword(user.password);
     const { password, ...userData } = user;
